Patrolling

+++ Chillenium/Assets/Scripts/TilemapStuff.cs
@@ -3,7 +3,7 @@ using UnityEngine.Tilemaps;
 
 public class TileTest : MonoBehaviour {
 
-    public byte[,] grid; // this is the array of things we own
+    public int[,] grid; // this is the array of things we own
     // assuming the map is a 5x5 and that the border is one tile wide
     
     void Start () {
@@ -12,7 +12,7 @@ public class TileTest : MonoBehaviour {
         BoundsInt bounds = tilemap.cellBounds;
         TileBase[] allTiles = tilemap.GetTilesBlock(bounds);
 
-        grid = new byte[bounds.size.x, bounds.size.y];
+        grid = new int[bounds.size.x, bounds.size.y];
         Debug.Log(bounds.size.x + " x " + bounds.size.y);
 
         for (int y = 0; y < bounds.size.y; y++) {
@@ -26,5 +26,14 @@ public class TileTest : MonoBehaviour {
                 }
             }
         }
+
+
+        // for(int y = 0; y < bounds.size.y; y++){
+        //     string row = "";
+        //     for(int x = 0; x < bounds.size.x; x++){
+        //         row += grid[x, y];
+        //     }
+        //     Debug.Log(row);
+        // }
     }   
 }
